+++ src/main/java/com/reservashoteis/dto/response/ReservationResponseDto.java
@@ -3,9 +3,9 @@ package com.reservashoteis.dto.response;
 import java.time.LocalDate;
 
 public record ReservationResponseDto(
-    int idReservation,
-    int idClient,
-    int idRoom,
+    Long idReservation,
+    Long idClient,
+    Long idRoom,
     LocalDate checkInDate,
     LocalDate checkOutDate,
     Float totalValue
